requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
google-generativeai
python-dotenv
